refactor: simplify heading shrinking
Takes advantage of Pandoc accepting an integer header level to shrink
headings in nested chapters using arithmetic instead of repeatedly
applying a shrink-by-one-HeadingLevel mapping. This also removes a noisy
log when headings are shrunk to levels smaller than 6.

+++ src/pandoc/native.rs
@@ -850,7 +850,7 @@ impl<'a, 'book, 'p, W: io::Write> SerializeBlock<'a, 'book, 'p, W> {
     /// Header - level (integer) and text (inlines)
     pub fn serialize_header(
         self,
-        level: usize,
+        level: u16,
         attrs: impl Attributes,
         inlines: impl FnOnce(&mut SerializeInlines<'a, 'book, 'p, W>) -> anyhow::Result<()>,
     ) -> anyhow::Result<()> {
